requirements.txt
cudatoolkit==10.0
keras-gpu==2.2.4
sklearn
ujson
gensim
pandas
hdf5==1.10.4
matplotlib-3.0.2
seaborn==0.9.0
nltk==3.4
